fix: delete policyを追加する

+++ app/Http/Controllers/Articles/PostArticles/PostArticleIdDeleteController.php
@@ -7,6 +7,7 @@ namespace App\Http\Controllers\Articles\PostArticles;
 use App\Http\Controllers\Controller;
 use App\Http\Controllers\GetIndexController;
 use App\Infrastructure\Persistence\PublishedArticleRepository;
+use App\Models\ArticlePublishedEloquent;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\DB;
 
@@ -18,6 +19,10 @@ class PostArticleIdDeleteController extends Controller
 
     public function __invoke(Request $request, string $article_id)
     {
+        if ($request->user()->cannot('delete', ArticlePublishedEloquent::query()->find($article_id))) {
+            abort(403);
+        }
+
         DB::transaction(function () use ($article_id) {
             $this->publishedArticleRepository->delete($article_id);
         });
